+++ discord/ext/vbu/cogs/utils/__init__.py
@@ -15,7 +15,7 @@ from .string import Formatter
 from .component_check import component_check, component_id_check
 from .embeddify import Embeddify
 from .twitch_stream import TwitchStream
-from .translation import translation, i8n
+from .translation import translation, i18n
 
 
 __all__ = (
@@ -54,7 +54,7 @@ __all__ = (
     'Redis',
     'Stats',
     'translation',
-    'i8n',
+    'i18n',
 )